+++ packages/cms/schemas/index.ts
@@ -1,0 +1,28 @@
+import { blogPostType } from "./blogPost";
+import { eventType } from "./event";
+import { pageType } from "./page";
+import { sermonType } from "./sermon";
+import { sermonSeriesType } from "./sermonSeries";
+import { siteConfigType } from "./siteConfig";
+import { staffBioType } from "./staffBio";
+
+/** All document types for Sanity Studio config. */
+export const schemaTypes = [
+  siteConfigType,
+  sermonType,
+  sermonSeriesType,
+  eventType,
+  blogPostType,
+  staffBioType,
+  pageType,
+];
+
+export {
+  blogPostType,
+  eventType,
+  pageType,
+  sermonType,
+  sermonSeriesType,
+  siteConfigType,
+  staffBioType,
+};